requirements.txt
asgiref==3.3.4
Django==3.2
django-markdownx==3.0.1
importlib-metadata==4.0.1
Markdown==3.3.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
